Move URL helpers to helpers.py. Rename some FAS urls.
